+++ lib/nanoc/base/source_data/item.rb
@@ -39,13 +39,6 @@ module Nanoc
     # @return [Nanoc::Site] The site this item belongs to
     attr_accessor :site
 
-    # @return [Nanoc::Item, nil] The parent item of this item. This can be
-    #   nil even for non-root items.
-    attr_accessor :parent
-
-    # @return [Array<Nanoc::Item>] The child items of this item
-    attr_accessor :children
-
     # Creates a new item with the given content or filename, attributes and
     # identifier.
     #
@@ -89,9 +82,6 @@ module Nanoc
       # TODO validate identifier (must start with slash, cannot end with slash)
       @attributes   = attributes.symbolize_keys_recursively
       @identifier   = identifier.freeze
-
-      @parent       = nil
-      @children     = []
 
       @reps         = []
     end
@@ -209,7 +199,6 @@ module Nanoc
     # @return [void]
     def freeze
       attributes.freeze_recursively
-      children.freeze
       identifier.freeze
       raw_filename.freeze if raw_filename
       raw_content.freeze  if raw_content